refactor: centralize theme api
